Update DbTable.php
The write() method returned false if the session data was not altered, which lead to a warning. This happens if two request for a session are fired at the same time (eg. file calls using the same session). 
Zend_Db_Adapter_Abstract->update() returns the number of affected rows and if the data does not change, it is 0.